add support for multi-prefabs, equipment items, and zeus
The prefab child field of a prefab can contain multiple prefabs so
support has been added to account for this and crawl all parallel
child prefabs recursively rather than just singular ones.

Support for Equipment items has been added to output info for things
like defuse kits, shields, bumpmines etc.

Explicit support for the Zeus has been added as this item, although
a weapon, is categorised as an Equipment item. It has been added
such that it will be treated as a skinnable weapon in the apps
output.